# client: small optimization to Batch.appendReqs