Add the requests directly to the Batch.reqs slice instead of creating a
temporary slice.

```
name                           old time/op    new time/op    delta
TrackChoices1_Cockroach-32        569µs ± 3%     573µs ± 1%     ~     (p=0.175 n=20+19)
TrackChoices10_Cockroach-32       145µs ± 2%     145µs ± 3%     ~     (p=0.414 n=20+20)
TrackChoices100_Cockroach-32     92.3µs ± 2%    91.9µs ± 3%     ~     (p=0.233 n=19+18)
TrackChoices1000_Cockroach-32    90.6µs ± 3%    86.6µs ± 3%   -4.32%  (p=0.000 n=17+19)

name                           old allocs/op  new allocs/op  delta
TrackChoices1_Cockroach-32          374 ± 0%       366 ± 0%   -2.14%  (p=0.000 n=18+16)
TrackChoices10_Cockroach-32         129 ± 0%       125 ± 0%   -3.10%  (p=0.000 n=16+20)
TrackChoices100_Cockroach-32        101 ± 0%        97 ± 0%   -3.96%  (p=0.000 n=20+20)
TrackChoices1000_Cockroach-32      98.0 ± 0%      94.0 ± 0%   -4.08%  (p=0.000 n=19+19)
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="35" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cockroachdb/cockroach/6785)
<!-- Reviewable:end -->